rebase decoupling_format
