+++ mathics/builtin/inout.py
@@ -19,7 +19,7 @@ from mathics.builtin.base import (
     Operator,
     Predefined,
 )
-from mathics.builtin.box.inout import RowBox, to_boxes, _BoxedString
+from mathics.builtin.box.inout import RowBox, to_boxes
 from mathics.builtin.comparison import expr_min
 from mathics.builtin.lists import list_boxes
 from mathics.builtin.options import options_to_rules
@@ -28,11 +28,12 @@ from mathics.builtin.tensors import get_dimensions
 from mathics.core.atoms import (
     Integer,
     Real,
-    MachineReal,
     PrecisionReal,
+    MachineReal,
     String,
     StringFromPython,
 )
+
 from mathics.core.attributes import (
     hold_all as A_HOLD_ALL,
     hold_all_complete as A_HOLD_ALL_COMPLETE,
@@ -40,9 +41,10 @@ from mathics.core.attributes import (
     protected as A_PROTECTED,
 )
 from mathics.core.convert.python import from_bool
-from mathics.core.element import EvalMixin
+from mathics.core.element import EvalMixin, BoxElement
 from mathics.core.expression import Expression, BoxError
 from mathics.core.evaluation import Message as EvaluationMessage
+from mathics.core.formatter import _BoxedString, format_element
 from mathics.core.list import ListExpression
 from mathics.core.number import (
     dps,
@@ -608,7 +610,7 @@ class MakeBoxes(Builtin):
     def apply_general(self, expr, f, evaluation):
         """MakeBoxes[expr_,
         f:TraditionalForm|StandardForm|OutputForm|InputForm|FullForm]"""
-        if isinstance(expr, BoxConstruct):
+        if isinstance(expr, BoxElement):
             expr = expr.to_expression()
         if isinstance(expr, Atom):
             return expr.atom_to_boxes(f, evaluation)
@@ -745,7 +747,7 @@ class ToBoxes(Builtin):
         form_name = form.get_name()
         if form_name is None:
             evaluation.message("ToBoxes", "boxfmt", form)
-        boxes = expr.format(evaluation, form_name)
+        boxes = format_element(expr, evaluation, form_name)
         return boxes
