+++ mathics/builtin/numbers/numbertheory.py
@@ -8,7 +8,7 @@ import mpmath
 import sympy
 
 from mathics.builtin.base import Builtin, SympyFunction
-from mathics.core.atoms import Integer, Integer0, Integer10, Rational, SymbolDivide
+from mathics.core.atoms import Integer, Integer0, Integer10, Rational
 from mathics.core.attributes import (
     listable as A_LISTABLE,
     numeric_function as A_NUMERIC_FUNCTION,
@@ -22,7 +22,7 @@ from mathics.core.convert.sympy import SympyPrime, from_sympy
 from mathics.core.evaluators import eval_N
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
-from mathics.core.symbols import Symbol, SymbolFalse
+from mathics.core.symbols import Symbol, SymbolDivide, SymbolFalse
 from mathics.core.systemsymbols import (
     SymbolCeiling,
     SymbolComplex,
