+++ mathics/builtin/string/operations.py
@@ -11,10 +11,6 @@ import zlib
 
 from mathics.algorithm.parts import python_seq, convert_seq
 
-from mathics.builtin.base import (
-    BinaryOperator,
-    Builtin,
-)
 
 from mathics.builtin.atomic.strings import (
     _StringFind,
@@ -24,6 +20,11 @@ from mathics.builtin.atomic.strings import (
     to_regex,
 )
 from mathics.builtin.box.inout import _BoxedString
+
+from mathics.builtin.base import (
+    BinaryOperator,
+    Builtin,
+)
 
 from mathics.core.atoms import (
     ByteArrayAtom,
@@ -40,6 +41,7 @@ from mathics.core.attributes import (
 )
 from mathics.core.convert.python import from_python
 from mathics.core.expression import Expression, string_list
+from mathics.core.formatter import format_element
 from mathics.core.list import ListExpression
 from mathics.core.symbols import (
     Symbol,
@@ -922,11 +924,9 @@ class StringRiffle(Builtin):
         # Getting all together
         result = left
         for i in range(len(liststr.elements)):
-            text = (
-                liststr.elements[i]
-                .format(evaluation, "System`OutputForm")
-                .boxes_to_text(evaluation=evaluation)
-            )
+            text = format_element(
+                liststr.elements[i], evaluation, "System`OutputForm"
+            ).boxes_to_text(evaluation=evaluation)
             if i == len(liststr.elements) - 1:
                 result += text + right
             else:
